Fixed several Book related issues (#416)
Closes #179
Closes #281
Closes #417

Built on the 'IBook' interface.
I made all the buttons that were private before, public and part of the api and I also added methods for each of the buttons expected funtions.
This now allows other systems to work with the book without the use of its buttons.
Its also keeping it flexible by allowing consumer code to decide what is supposed to happen when a button is pressed because they can override the 'OnClick' action of each button.
Right now the LauncherButton is part of IBook but when that is not needed anymore when the AugmentBook is introduced you could easily subclass Book or make another interface and move LauncherButton along with Title and LauncherParent in there and you're good. Or so I think anyway.

All the LauncherButton code in the ArtifactManager could be moved to Book theoretically but I decided to leave it for now to not screw up the desync issues fixed in in #372.