Implement DestructibleHider from the live vJASS version of the map (#576)
What this boils down to is spliting the entire map into sub-grids or 'tiles' asigning the tile an Id that can be derived from any coordinate from within the tile. Now all destructables on the map can be mapped to a tileId based on their coordinates.

We add desructables to a list that is mapped to their tileId in a dictionary. Desructables are loaded and unloaded based on where the palyer camera is by accessing the tile dictionary.