Elimination System
Implemented an elimination system that starting from turn 20 at the end of a game turn checks each player to see if they meet the elimination threshold. Any player that meets the  elimination threshold for 5 turns in a row get their score status set to defeated.

Added Essential flag to legend class, defaults to false on creation

Set Essential flag to true for legends that factions need to stay in the game