Player Distribution Queue (#2042)
* Victory Imminent Message Spam Fix

Get the former owner team of the changing control point and check if the new team is different from the former before doing victory checking.

Fixed an index out of bounds exception when trying to distribute units to other players in the faction

* Fix for Issue #46

Fix for issue of quest objectives not reflecting unit team changes

* Update ObjectiveControlPoint.cs

Removed unused code

* Goblin Zandalar desync fixes

Changed ZanadalarGoblinChoice.cs to be mostly static which mirrors how OpenAllianceVote.cs is done

Updated power book to update the tracked faction for a book regardless of if the player is the local one. This should keep clients in sync better

* Persistent Soul fix

Made RemoveAllAbilities() the rist action that happens to the reanimated unit.

SetTimedLife() adds a buff which was being removed by RemoveAllAbilities() and basically 'despelling' the summoned unit.

* Reveal Map During Cinematic

When intro text starts reveal map and when it ends hide map again

* CP objectives and CP events

Added player and team events for when control points are added or removed
Added new faction and team objectives for control points being less than a given amount
Added new faction and team objectives for control points being greater than a given amount
Added Team defeat objective

* RemoveControlPoint

Change to use correct method RemoveControlPoint()

* Team CP Rework

Changed to loop over all members of the team and get the individual control points rather than keep a separate list updated

* Fixed Typo

Fixed Typo

* Quest Progress Reset

Added resetting of quest objectives

* On faction join quest fix

If owning player is null then return null. This ensures QuestProgress get set to Incomplete

* Update QuestExplosiveEngineering.cs

Display object name in reward description rather than objectid

* Player Queue

Add players to a queue when they leave so that unit distribution can happen one player at a time.

Added a function to the capital manager to see if a unit is a capital or not