Removed trigger string localization (#2069)
In the WC3 file format, strings (like a unit's name) actually point to a TriggerString dictionary which gets created on map save, for the purposes of localization. This is useless to us since we don't do any language localization. This change gets rid of the TriggerStrings concept entirely, opting to instead store any text values directly in the object data. This saves a tone of file space and makes our build process more deterministic.