Dialogue Presenter Changes (#2437)
* Victory Imminent Message Spam Fix

Get the former owner team of the changing control point and check if the new team is different from the former before doing victory checking.

Fixed an index out of bounds exception when trying to distribute units to other players in the faction

* Fix for Issue #46

Fix for issue of quest objectives not reflecting unit team changes

* Update ObjectiveControlPoint.cs

Removed unused code

* Goblin Zandalar desync fixes

Changed ZanadalarGoblinChoice.cs to be mostly static which mirrors how OpenAllianceVote.cs is done

Updated power book to update the tracked faction for a book regardless of if the player is the local one. This should keep clients in sync better

* Persistent Soul fix

Made RemoveAllAbilities() the rist action that happens to the reanimated unit.

SetTimedLife() adds a buff which was being removed by RemoveAllAbilities() and basically 'despelling' the summoned unit.

* Reveal Map During Cinematic

When intro text starts reveal map and when it ends hide map again

* CP objectives and CP events

Added player and team events for when control points are added or removed
Added new faction and team objectives for control points being less than a given amount
Added new faction and team objectives for control points being greater than a given amount
Added Team defeat objective

* RemoveControlPoint

Change to use correct method RemoveControlPoint()

* Team CP Rework

Changed to loop over all members of the team and get the individual control points rather than keep a separate list updated

* Fixed Typo

Fixed Typo

* Quest Progress Reset

Added resetting of quest objectives

* On faction join quest fix

If owning player is null then return null. This ensures QuestProgress get set to Incomplete

* Update QuestExplosiveEngineering.cs

Display object name in reward description rather than objectid

* Player Queue

Add players to a queue when they leave so that unit distribution can happen one player at a time.

Added a function to the capital manager to see if a unit is a capital or not

* Don't remove protectors

Added UnitIsProtector method which checks if the unit is a protector of any capital

Added UnitIsCapitalProtector checks if unit is a protector of a specific capital

Don't remove a unit that is a protector when a player leaves the game

* Don't remove portals

Added new unitType category of 'Portal'
Setup units that need to be categorised as 'Portal'
Don't remove units that are category 'Portal' when a player leaves

* Use Meta flag

Updated portal units to be flag as 'Meta' that way they should never be deleted

* Change Meta to NeverDelete

Changed 'Meta' to 'NeverDelete' so that it is more clear what that flag represents with out having to go to the comments in the definition

* Never Delete Traders

Removed 'Portal' category as not used
Set traders to NeverDelete

* Fixed Errors with UnitTypeConfig

* Unit distribution

Instead of just removing a unit during unit distribution Kill() it as well. Doing this will mean that death events still get triggered.

Set all gate types to 'NeverDelete'

* Settings Save/Load System

Updated WCSharp.Json and WCSharp.Shared so that WCSharp.SaveLoad could be added.

Added a save and load system using WCSharp.SaveLoad for player settings.

Added commands for toggling Dialogue, QuestText and Captions

Added a command to see current settings

* Tidy up

Updated descriptions of new commands
Made PlayerData an internal class again

* Tooltip Tidy

Fixed wrong comments
Made tooltips clearer

* Fix merge

I think I messed up resolving a merge conflict so this is a fix for that.

* Dialogue Presenter Changes

Reworked the dialogue presenter to be more extendable.
Added an "invasion location" choice dialogue for scourge to pick when completing the plague quest

* update WCSharp

---------

Co-authored-by: Sebastian Tanoni <47048283+SebastianTanoni@users.noreply.github.com>