Refactor Factions to allow for optional and late registration (#2594)
Currently, all Factions must be set up at the start of the game so that their associated quest setups can be run as well.
This refactor allows Factions to be registered at any time. It bundles relevant initialization with the Faction, rendering that initialization optional. This will become more and more important the more optional Factions we add, as we don't want to run Faction initialization for Factions that do not and never will exist in a given game.