Refactor Faction initialization to support Faction-dependent initializers (#2705)
The faction setup system was built with the assumption that all factions should be registered in every game. With faction choices, this is no longer true. The system has to evolve to support the idea that any number and combination of factions might be registered in a given session, and at any time of the game.

This poses a significant problem for initialization steps that depend on the existence of another faction. For instance, how do you initialize a dialogue sequence between Jaina and Archimonde without knowing that they both exist?

To solve this problem I've implemented FactionDependentInitializers. They can be registered in a Factions constructor via the RegisterFactionDependentInitializer<TFaction> generic method. For example, one could write the following code in Legion's constructor:
```RegisterFactionDependentInitializer<Dalaran>(RegisterDialogue);```
The function above tells the system that, when both Legion and Dalaran have been successfully registered, it should run the RegisterDialogue function. If either Legion or Dalaran are not registered, the function will never run.

Currently this system is only used for Dialogue, but in the future it could be used for setting up Quests that relate to one or more optional factions.