Remove fallback Faction color system
A while ago we added a (very well-written!) fallback Faction color system so that Factions with overlapping ideal lore-related colors could simultaneously attempt to claim those colors, with one Faction instead falling back to a pre-specified fallback color.

On reflection, it will be a lot clearer for the game if we can just tie color to a particular faction slot. So, practically:
- Gilneas is now pink
- Sunfury is now violet
- Warsong is now red
- Draenei are now mint