Improve performance of passive abilities by conditionally registering events (#3383)
Prior to this change, a large number of events were naively registered for every Passive Ability registration, independently of whether that Passive Ability actually needed it.
With this change, Passive Abilities must implement one or more of the provided effect interfaces to be registered to a particular event. This should, hopefully, improve performance.